+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Widget "AI assistants" - requires Matomo 5.5+ (#64)
+
 ## [3.3.0] - 2025-09-23
 
 ### Added